pyproject.toml
[tool.poetry]
name = "vcsws"
version = "0.1.0"
description = "Version Control System Using WebSockets"
authors = ["Maksim Shushkevich <meshushkevich@edu.hse.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^12.0"
tqdm = "^4.66.4"
pystun3 = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
